POC: start getting rid of parameters
This commit gets rid of the defaults.conf file and the 'parameters' dict
that was used to handle the configuration of bots. It instead uses class
variables and the defaults are set in the relevant parts of the code,
but can be easily overriden using environment variables or configuration
file.
This is still early in the process and there is still a lot of code that
has to be updated, both to replace 'parameters' and also stylistic
(replace the call to getattr, that was used for getting values from
'parameters')